imageembed/requirements.txt
aiohttp==3.8.5
aiomysql==0.1.1
aiopg==1.4.0
aiosignal==1.3.1
amqp==5.1.1
anyio==3.6.1
appdirs==1.4.3
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
Automat==0.8.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.4
bytecode==0.15.1
cattrs==23.1.2
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==20.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
confluent-kafka==2.2.0
constantly==15.1.0
cryptography==38.0.1
cssselect==1.2.0
databases==0.6.1
dbus-python==1.2.16
ddsketch==2.0.4
ddtrace==2.1.0
Deprecated==1.2.13
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.2.1
elasticsearch==7.17.9
email-validator==1.3.0
entrypoints==0.3
envier==0.4.0
exceptiongroup==1.1.3
fastapi==0.85.0
fastapi-limiter==0.1.5
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.12.2
greenlet==1.1.3
h11==0.14.0
httplib2==0.14.0
httptools==0.5.0
huggingface-hub==0.19.4
hyperlink==19.0.0
idna==3.4
importlib-metadata==6.8.0
incremental==16.10.1
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libvirt-python==6.1.0
lxml==4.9.3
MarkupSafe==2.1.1
minio==7.1.12
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.4
netifaces==0.10.4
networkx==3.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.1
oauthlib==3.1.0
opentelemetry-api==1.20.0
orjson==3.8.0
packaging==21.3
pandas==2.1.2
Pillow==10.1.0
prompt-toolkit==3.0.31
protobuf==4.24.4
psutil==5.9.5
psycopg2-binary==2.9.9
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.5.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
regex==2023.10.3
requests==2.28.1
requests-file==1.5.1
requests-unixsocket==0.2.0
safetensors==0.4.1
schedule==1.1.0
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==2.3.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==18.1.0
sgmllib3k==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.41
ssh-import-id==5.10
starlette==0.20.4
sympy==1.12
systemd-python==234
threadpoolctl==3.2.0
tinysegmenter==0.3
tldextract==5.0.1
tokenizers==0.15.0
torch==2.1.2+cpu
torchaudio==2.1.2+cpu
torchvision==0.16.2+cpu
tqdm==4.66.1
transformers==4.36.2
Twisted==18.9.0
typing_extensions==4.3.0
tzdata==2023.3
ubuntu-advantage-tools==20.3
ufw==0.36
ujson==5.5.0
unattended-upgrades==0.1
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.17.0
vine==5.0.0
virtualenv==20.0.17
wadllib==1.3.3
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.15.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.17.0
zope.interface==4.7.1
